+++ autd3-driver/src/datagram/stm/gain.rs
@@ -27,7 +27,7 @@ impl<G: Gain> GainSTM<G> {
     ///
     /// * `freq` - Frequency of STM. The frequency closest to `freq` from the possible frequencies is set.
     ///
-    pub const fn from_freq(freq: float) -> Self {
+    pub const fn from_freq(freq: f64) -> Self {
         Self::from_props_mode(STMProps::from_freq(freq), GainSTMMode::PhaseIntensityFull)
     }
 
@@ -57,7 +57,7 @@ impl<G: Gain> GainSTM<G> {
         )
     }
 
-    pub fn frequency(&self) -> float {
+    pub fn frequency(&self) -> f64 {
         self.props.freq(self.gains.len())
     }
 
@@ -181,10 +181,10 @@ mod tests {
     #[case(0.5, 2)]
     #[case(1.0, 10)]
     #[case(2.0, 10)]
-    fn test_from_requency(#[case] freq: float, #[case] n: usize) -> anyhow::Result<()> {
+    fn test_from_requency(#[case] freq: f64, #[case] n: usize) -> anyhow::Result<()> {
         let stm = GainSTM::from_freq(freq).add_gains_from_iter((0..n).map(|_| NullGain {}))?;
         assert_eq!(freq, stm.frequency());
-        assert_eq!(freq * n as float, stm.sampling_config()?.frequency());
+        assert_eq!(freq * n as f64, stm.sampling_config()?.frequency());
         Ok(())
     }
