+++ autd3/benches/gain/helper.rs
@@ -6,8 +6,8 @@ pub fn generate_geometry(size: usize) -> Geometry {
             .flat_map(|i| {
                 (0..size).map(move |j| {
                     AUTD3::new(Vector3::new(
-                        i as float * AUTD3::DEVICE_WIDTH,
-                        j as float * AUTD3::DEVICE_HEIGHT,
+                        i as f64 * AUTD3::DEVICE_WIDTH,
+                        j as f64 * AUTD3::DEVICE_HEIGHT,
                         0.,
                     ))
                     .into_device(j + i * size)
