+++ examples/src/tests/stm.rs
@@ -8,7 +8,7 @@ pub async fn focus_stm(autd: &mut Controller<impl Link>) -> anyhow::Result<bool>
     let point_num = 200;
     let radius = 30.0 * MILLIMETER;
     let stm = FocusSTM::from_freq(1.0).add_foci_from_iter((0..point_num).map(|i| {
-        let theta = 2.0 * PI * i as float / point_num as float;
+        let theta = 2.0 * PI * i as f64 / point_num as f64;
         let p = radius * Vector3::new(theta.cos(), theta.sin(), 0.0);
         center + p
     }))?;
@@ -29,7 +29,7 @@ pub async fn gain_stm(autd: &mut Controller<impl Link>) -> anyhow::Result<bool> 
     let radius = 30.0 * MILLIMETER;
 
     let stm = GainSTM::from_freq(1.0).add_gains_from_iter((0..point_num).map(|i| {
-        let theta = 2.0 * PI * i as float / point_num as float;
+        let theta = 2.0 * PI * i as f64 / point_num as f64;
         let p = radius * Vector3::new(theta.cos(), theta.sin(), 0.0);
         Focus::new(center + p)
     }))?;
